refactor(scalars): move OIDInput to ui
