+++ packages/design-system/src/scalars/components/oid-field/index.ts
@@ -1,2 +1,2 @@
 export { OIDField } from "./oid-field.js";
-export type { OIDFieldProps, OIDOption } from "./types.js";
+export type { OIDFieldProps } from "./types.js";
