+++ packages/design-system/src/ui/components/data-entry/oid-input/index.ts
@@ -1,0 +1,2 @@
+export { OIDInput } from "./oid-input.js";
+export type { OIDInputProps, OIDOption } from "./types.js";
